ApiTypes

Description

Types used in the BugWalker API

Documentation

data BuildT

Api representation of a GameData build

Constructors

BuildT

Fields

buildIdT :: Int

buildStringT :: String

buildDateT :: String

data SpellT

Api representation of a GameData Spell

Constructors

SpellT

Fields

spellIdT :: Int

spellNameT :: String

spellDescT :: String

spellBuildT :: String

data Bug

Primary type of a single Bug

Constructors

Bug

Fields

bugId :: Int64

bugDateModified :: String

bugSeverity :: BugSeverity

bugType :: BugType

bugTitle :: String

bugStatus :: BugStatus

bugTags :: Maybe String

bugSpellId :: Int

bugBuildId :: Int

bugSpec :: Specialisation

bugBlueTrackerLink :: Maybe String

bugSpellName :: Maybe String

bugUserName :: Maybe String

bugUserId :: Int64

bugBuildString :: Maybe String

data BugDetail

Detailed representation of a Bug including the Content fields

Constructors

BugDetail

Fields

bugdtId :: Int64

bugdtDateCreated :: String

bugdtDateModified :: Maybe String

bugdtSeverity :: BugSeverity

bugdtType :: BugType

bugdtTitle :: String

bugdtStatus :: BugStatus

bugdtTags :: Maybe String

bugdtSpellId :: Int

bugdtBuildId :: Int

bugdtSpec :: Specialisation

bugdtDescription :: Maybe String

bugdtSteps :: Maybe String

bugdtContent :: Maybe String

bugdtBlueTrackerLink :: Maybe String

bugdtSubmitter :: Int64

data BugSeverity

Represents the severity levels that a Bug can have

Constructors

Low

Medium

Critical

data BugType

Represents the different types that a Bug can have

Constructors

Visual

Mechanical

Gameplay

System

data Specialisation

Represents the possible Specilisations that a monk can have

Constructors

All

Windwalker

Brewmaster

Mistweaver

data BugStatus

Represents the statuses that a Bug can have

Constructors

Closed

Open

Verified

Archived

data ResponseMessage

Api general ResponseMessage

Constructors

ResponseMessage

Fields

resMessageStatusCode :: Int

resMessageStatus :: String

resMessage :: String

data LoginResponseMessage

Api response type to a login request

Constructors

LoginResponseMessage

Fields

lrmStatusCode :: Int

lrmMessage :: String

lrmToken :: String

data BugChangedResponseMessage

Api response type to a Bug insertion, delete or update

Constructors

BugChangedResponseMessage

Fields

bcMessageStatusCode :: Int

bcMessageBugId :: Int64

bcMessageStatus :: String

bcMessage :: String

data UserChangedResponseMessage

Api response type to a User insertion, delete or update

Constructors

UserChangedResponseMessage

Fields

ucMessageStatusCode :: Int

ucMessageUserId :: Int64

ucMessageStatus :: String

ucMessage :: String

data Comment

Representation of a single Comment

Constructors

Comment

Fields

cId :: Int64

cCreatedTime :: String

cModifiedTime :: String

cUserId :: Int64

cUserName :: String

cUserRole :: Int

cUserAvatar :: Int

cBugId :: Int64

cStatus :: Int

cBody :: String