GameData

Description

The main gamedata module which exports functions for running a full update of the gamedata by fetching data from Github, parsing it and inserting it to a given db

Documentation

runGameDataUpdate

Arguments

:: Bool
should do hardInit of db

-> Bool
fetch from local file (true) or via http (false)

-> IO ()

Main entry point of the module, attempts to do a full gamedata update