IOH

Description

IO Abstraction handler used to interact with an underlying data store. All Datastore io should go through this.

Documentation

initIO

Arguments

:: Bool
Do a hardInit

-> IO ()

Calls on the underlying provider to init its datastore

storeSpellData

Arguments

:: SpellData
the SpellData to store

-> IO ()

Attempt to store a full SpellData collection (from a GameData update)

getBuilds :: IO [BuildT]

Fetches all builds from the data store

getSpells :: IO [SpellT]

Fetches all spells from the data store

storeBug

Arguments

:: BugDetail
The bug to insert

-> IO (Maybe Int64)
IO (id of the inserted bug or Nothing)

Attempt to insert a Bug into the data store

changeBug

Arguments

:: AuthenticatedUser
the user making the request

-> Int64
the unique id of the bug to update

-> BugDetail
the modified bug detail data

-> IO (Either IOHError Int64)

Attempt to update the stored data of a bug

getBug

Arguments

:: Int64
Unique Id of the bug to fetch

-> IO (Maybe BugDetail)

Attempt to fetch a bug from the data store

removeBug

Arguments

:: AuthenticatedUser
The user making the request

-> Int64
the unique id of the bug to delete

-> IO (Either IOHError Bool)

Attempt to delete a bug from the data store

getBugs

Arguments

:: Int
Limit on how many bugs to fetch

-> IO [Bug]

Fetches all the bugs from the data store

getBugsByUser

Arguments

:: Int64
the userId to fetch submitted bugs for

-> IO [Bug]

Fetches all the bugs from the data store that were submitted by the provided userId

storeUser

Arguments

:: User
the user to insert

-> IO (Either IOHError Int64)

Attempts to insert a User in to the data store

getUser

Arguments

:: Int64
unique UserId of the user to fetch

-> IO (Either IOHError User)

Attempts to fetch a user from the data store

getUserByName

Arguments

:: String
unique username of the user to fetch

-> IO (Either IOHError User)

Attempts to fetch a user from the data store (by name)

getUsers :: IO [UserL]

Fetches all users fromthe data store

getUserL

Arguments

:: Int64
unique userId of the user to fetch

-> IO (Either IOHError UserL)

Fetches a subset of User properties from the data store (for User listing)

changeUserRole

Arguments

:: AuthenticatedUser
The requesting user (Mod Only)

-> Int64
the unique userId of the user to change role for

-> Int
the userRole to change to

-> IO (Either IOHError Int64)

(Mod only) Attempts to change the userRole of a given user

storeComment

Arguments

:: AuthenticatedUser
the requesting user

-> Comment
the comment to insert

-> IO (Either IOHError Int64)

Attempts to insert a Comment into the data store

getComments

Arguments

:: Int64
the unique bugId to fetch comments for

-> IO (Either IOHError [Comment])

Fetches all comments for a given bugId

deleteComment

Arguments

:: AuthenticatedUser
the requesting user

-> Int64
The id of the comment to delete

-> IO (Either IOHError Bool)

Deletes a single comment from the data store

data IOHError

Represents differnet errors that can occur during IOH actions

Constructors

IOHInsertError String

IOHFetchError String

IOHDeleteError String

IOHAuthError String

IOHUpdateError String