JwkStore

Description

Functions related to storing and extracting the master JWK to disk

Documentation

loadJwk

Arguments

:: FilePath
the location to readwrite fromto

-> IO (Bool, JWK)
(Did read from disk, the JWK)

Attempts to either load an existing JWK from store (disk) or generates a new JWK and returns it in a tuple with the first bool indicating whether or not it was read or generated