Logging

Description

Provides threadsafe logging using a TQueue, should run on a dedicated logging thread

Documentation

startLoggingThread

Arguments

:: LogLevel
The minimum log level for messages to be logged

-> FilePath
path of the file to log to

-> ReadQueue LogEntry
readqueue to dequeue entries from

-> IO ()

The main logging thread entry point, should be run in forkIO

data LogEntry

Constructors

LogEntry

Fields

logEntryLevel :: LogLevel

logEntryTag :: Text

logEntryText :: Text

data Logger

The logging context to be passed around for enqueuing log entries

Constructors

Logger

Fields

lcLogQueue :: !(WriteQueue LogEntry)

data LogLevel

Constructors

Debug

Info

Warning

Error

All

logDebug

Arguments

:: Loggable l

=> l
The logger to use

-> Text
The sender or module

-> Text
The message to log

-> IO ()

Log a debug message

logError

Arguments

:: Loggable l

=> l
The logger to use

-> Text
The sender or module

-> Text
The message to log

-> IO ()

Log an error message

logWarning

Arguments

:: Loggable l

=> l
The logger to use

-> Text
The sender or module

-> Text
The message to log

-> IO ()

Log a warning message

logInfo

Arguments

:: Loggable l

=> l
The logger to use

-> Text
The sender or module

-> Text
The message to log

-> IO ()

Log an info message