Parser

Description

Contains functions needed for parsing spelldump data into ADTs.

Documentation

parseSpells

Arguments

:: String
Delimiter to use when splitting SpellData dump

-> [String]
The SpellData Dump chunks

-> [Spell]

Attempts to parse a list of SpellData dump chunks to extract a list of Spells.

parseSpellData

Arguments

:: String
Delimiter to use when splitting SpellData dump

-> String
Delimiter to use when splitting the individual Spell data

-> String
the SpellData dump

-> Maybe SpellData

Attempts to parse a SpellData dump and generate the full SpellData representation of it