SQLiteFactory

Description

Contains functions for creating the underlying SQLite datastore.

Documentation

createDb

Arguments

:: Bool
Whether or not asset tables should be dropped

-> IO ()

Creates a new db or connects to an existing one at the default path and creates the basic table structure

dbPath :: String

The default path of the sqlite data store

Orphan instances

FromRow Bug

Instance details

Methods

fromRow :: RowParser Bug

FromRow BugDetail

Instance details

Methods

fromRow :: RowParser BugDetail

FromRow SpellT

Instance details

Methods

fromRow :: RowParser SpellT

FromRow Comment

Instance details

Methods

fromRow :: RowParser Comment

FromRow UserL

Instance details

Methods

fromRow :: RowParser UserL

FromRow User

Instance details

Methods

fromRow :: RowParser User

FromField UUID

Instance details

Methods

fromField :: FieldParser UUID

FromField Specialisation

Instance details

Methods

fromField :: FieldParser Specialisation

FromField BugStatus

Instance details

Methods

fromField :: FieldParser BugStatus

FromField BugSeverity

Instance details

Methods

fromField :: FieldParser BugSeverity

FromField BugType

Instance details

Methods

fromField :: FieldParser BugType

ToRow User

Instance details

Methods

toRow :: User -> [SQLData]