SQLiteProvider

Description

Contains functions for interacting with the underlying SQLite datastore.

Documentation

createDb

Arguments

:: Bool
Whether or not asset tables should be dropped

-> IO ()

Creates a new db or connects to an existing one at the default path and creates the basic table structure

insertSpellData :: SpellData -> IO ()

selectBuilds :: IO [(Int, String, String)]

selectSpells :: IO [SpellT]

insertBug

Arguments

:: Int
the severity of the bug

-> Int
the type of the bug

-> String
the title of the bug

-> String
the tags of the bug

-> Int
the spellId of the bug

-> Int
the buildId of the bug

-> Int
the Specilalisation of the bug

-> String
the description of the bug

-> String
the steps of the bug

-> String
the content of the bug

-> String
the blueTrackerLink of the bug

-> Int64
the userId that created the bug

-> IO (Maybe Int64)
On success: The id that the bug was assigned in the db

Inserts a single bug into the database @todo refactor to use ToRow

selectBug :: Int64 -> IO (Maybe BugDetail)

updateBug :: Int64 -> BugDetail -> IO (Maybe Int64)

deleteBug :: Int64 -> IO Bool

selectBugs :: Int -> IO [Bug]

selectBugs' :: Int -> IO [Bug]

insertUser :: User -> IO (Maybe Int64)

selectUser :: Int64 -> IO (Maybe User)

selectUserL :: Int64 -> IO (Maybe UserL)

selectUserByName :: String -> IO (Maybe User)

selectBugsByUser :: Int64 -> IO [Bug]

selectUsers :: IO [UserL]

updateUserRole :: Int64 -> Int -> IO Bool

insertComment :: Comment -> IO (Maybe Int64)

selectComment

Arguments

:: Int64
the id of the comment to fetch

-> IO (Maybe Comment)

Attempts to fetch a single comment from the data store

selectComments :: Int64 -> IO [Comment]

deleteComment

Arguments

:: Int64
the id of the comment to delete

-> IO Bool

Deletes a single comment from the data store

deleteCommentsByUser

Arguments

:: Int64
the id of the user to delete comments for

-> IO Bool

Deletes all comments submitted by a User (used when deleting the user)