UserApi

Description

Handlers related to User API endpoints

Documentation

userRegistrationHandler

Arguments

:: UserRegistration
UserRegistration object to store

-> Handler UserChangedResponseMessage

Handles user registration requests by asking the AuthHandler to create the user and then asks the IOH to store the new user

userLoginHandler

Arguments

:: JWTSettings
the JWTSettings

-> UserLoginData
the UserLoginData object

-> Handler LoginResponseMessage

Handles user login by getting the User object from the IOH and asking the AuthHandler to authenticate. Creates a JWT token on success

getUsersHandler :: Handler [UserL]

Handles listing users

getUserPHandler

Arguments

:: Int64
UserId of the user to fech profile for

-> Handler UserP

Handles feching of UserProfile objects

changeUserRoleHandler

Arguments

:: AuthenticatedUser
The user making the request

-> UserChangeRoleData
The user change data to make

-> Handler UserChangedResponseMessage

Handles changing a user's role (mod only)